MSVC appeasement.
* Changed over more wonky casts.
* MSVC warnings are very VERY picky (not a bad thing normally, but this is a
  90's code base ;p), so we're only using /W1 /WX there.
* Downgraded required cmake version to the one bundled in VS2022.
* Finally implemented file calls that work in Win32.

+++ src/render_soft/soft_r_draw.c
@@ -326,7 +326,7 @@ R_EmitCachedEdge
 void R_EmitCachedEdge(void) {
     edge_t *pedge_t;
 
-    pedge_t = (edge_t *)((unsigned long)r_edges + r_pedge->cachededgeoffset);
+    pedge_t = (edge_t *)((uintptr_t)r_edges + r_pedge->cachededgeoffset);
 
     if(!pedge_t->surfs[0]) {
         pedge_t->surfs[0] = surface_p - surfaces;
@@ -401,8 +401,8 @@ void R_RenderFace(msurface_t *fa, int clipflags) {
                         continue;
                     }
                 } else {
-                    if((((unsigned long)edge_p - (unsigned long)r_edges) > r_pedge->cachededgeoffset) &&
-                       (((edge_t *)((unsigned long)r_edges + r_pedge->cachededgeoffset))->owner == r_pedge)) {
+                    if((((uintptr_t)edge_p - (uintptr_t)r_edges) > r_pedge->cachededgeoffset) &&
+                       (((edge_t *)((uintptr_t)r_edges + r_pedge->cachededgeoffset))->owner == r_pedge)) {
                         R_EmitCachedEdge();
                         r_lastvertvalid = false;
                         continue;
@@ -436,8 +436,8 @@ void R_RenderFace(msurface_t *fa, int clipflags) {
                 } else {
                     // it's cached if the cached edge is valid and is owned
                     // by this medge_t
-                    if((((unsigned long)edge_p - (unsigned long)r_edges) > r_pedge->cachededgeoffset) &&
-                       (((edge_t *)((unsigned long)r_edges + r_pedge->cachededgeoffset))->owner == r_pedge)) {
+                    if((((uintptr_t)edge_p - (uintptr_t)r_edges) > r_pedge->cachededgeoffset) &&
+                       (((edge_t *)((uintptr_t)r_edges + r_pedge->cachededgeoffset))->owner == r_pedge)) {
                         R_EmitCachedEdge();
                         r_lastvertvalid = false;
                         continue;
